Take __TEXT vmaddr into account.
When looking up addresses in the mach-o object for an executable (not a library),
the following is true:
 - The addresses we look up are relative to the start of the __TEXT segment (they're small numbers).
 - The addresses reported by symbol.address() are also relative to the __TEXT segment.
 - The vmaddr of the __TEXT segment is usually 4GB.
 - The addresses in the debug info (low_pc, high_pc, etc.) are vm addrs; they're big numbers and
   include the 4GB offset.

So when we look up debug info for addresses, we need to add this 4GB offset to the addresses we look up.

+++ lib/src/dwarf.rs
@@ -8,14 +8,14 @@ use std::borrow::Cow;
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct AddressPair {
     pub original_address: u32,
-    pub address_in_this_object: u32,
+    pub address_in_this_object: u64,
 }
 
 impl AddressPair {
     pub fn same(address: u32) -> Self {
         AddressPair {
             original_address: address,
-            address_in_this_object: address,
+            address_in_this_object: address as u64,
         }
     }
 }
